+++ src/types/Constants.ts
@@ -71,7 +71,8 @@ export enum PAGES {
   APPOVERVIEW = 'appoverview',
   APPOVERVIEW_NEW = 'appoverview-new',
   DEACTIVATE = 'deactivate',
-  CHANGEIMAGE = 'changeimage',
+  CHANGE_IMAGE = 'changeimage',
+  CHANGE_DESCRIPTION = 'changedescription',
   APPRELEASEPROCESS = 'appreleaseprocess',
   APP_RELEASE_PROCESS_FORM = 'appreleaseprocess_form',
   INTRODUCTION = 'companyroles',
@@ -94,6 +95,7 @@ export enum PAGES {
   USECASE_PARTICIPATION = 'usecase-participation',
   ABOUTPAGE = 'about',
   CERTIFICATE_CREDENTIAL = 'certificate-credential',
+  DATA_SPACE = 'dataspace',
 }
 
 export enum OVERLAYS {